gemini_ocr_pdf_image/pyproject.toml
[project]
name = "gemini-ocr-pdf-image"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.19.0",
    "pillow>=11.2.1",
    "pymupdf>=1.26.0",
    "tqdm>=4.67.1",
    "python-dotenv>=1.0.0",
]
